+++ data-raw/cw_issue_body.R
@@ -1,0 +1,27 @@
+cw_issue_body <- "**Prepare**
+- [ ] Find Co-host and decide on theme
+- [ ] Create Event
+- [ ] Set up 1-week before posts
+  - [ ] Mastodon
+  - [ ] Slack (#general / #co-working)
+
+**Week of/before**
+- [ ] Contact co-host to create intro slide
+- [ ] Update slides to current theme (title slide, intro slide, and final slide)
+- [ ] Create coworking document
+  - [ ] Update doc and make **editable**
+  - [ ] Add link on main docs page
+- [ ] Set up 1-hour before posts referring to 1-week before posts
+  - [ ] Mastodon
+  - [ ] Slack (#general)
+- [ ] Run Coworking :tada: (Share link to coworking document)
+
+**Wrap up**
+- [ ] Add outstanding details to document
+- [ ] Make document read only
+- [ ] Add document link to event page
+- [ ] Post summary on Staff Slack
+- [ ] Post summary with link to next event on Mastodon"
+
+
+usethis::use_data(cw_issue_body, overwrite = TRUE)